### 2021-12-29 #4028 @whedon
[PRE REVIEW]: mixR: An R package for Finite Mixture Modeling for Both Raw and Binned Data
**Submitting author:** @GaryBAYLOR (<a href="http://orcid.org/0000-0003-0519-9605">Youjiao Yu</a>)
**Repository:** <a href="https://github.com/GaryBAYLOR/mixR.git" target ="_blank">https://github.com/GaryBAYLOR/mixR.git</a>
**Version:** 0.2.0
