### 2021-12-29 #4028 @arfon
@GaryBAYLOR – many thanks for your submission to JOSS. While we look for an editor, if you want to fix the DOIs (noting that whedon's suggestions are not always right), you can, then use the following commands (one at a time, as the first line of a new comment) to regenerate the PDF and check the references.
