+++ paper/paper.bib
@@ -27,6 +27,7 @@
   number={3},
   pages={609--617},
   year={1996},
+  doi={10.1111/j.2517-6161.1996.tb02104.x},
   publisher={Wiley Online Library}
 }
 
@@ -38,6 +39,7 @@
   number={3},
   pages={318--324},
   year={1987},
+  doi={10.2307/2347790},
   publisher={Wiley Online Library}
 }
 
@@ -49,6 +51,7 @@
   number={18},
   pages={4550--4568},
   year={2019},
+  doi={10.1080/03610926.2018.1494838},
   publisher={Taylor \& Francis}
 }
 
@@ -69,12 +72,14 @@
   number={1},
   pages={289},
   year={2016},
+  doi={10.32614/rj-2016-021},
   publisher={NIH Public Access}
 }
 
 @article{flexmix,
   title={Flexmix: A general framework for finite mixture models and latent glass regression in R},
   author={Leisch, Friedrich},
+  doi={10.18637/jss.v011.i08},
   year={2004}
 }
 
@@ -124,6 +129,7 @@
   number={4},
   pages={1--17},
   year={2020},
+  doi={10.1007/s10661-020-8223-4},
   publisher={Springer}
 }
 
@@ -134,6 +140,7 @@
   volume={170},
   pages={115369},
   year={2020},
+  doi={10.1016/j.watres.2019.115369},
   publisher={Elsevier}
 }
 
@@ -143,6 +150,7 @@
   journal={Journal of Sustainable Forestry},
   pages={1--15},
   year={2020},
+  doi={10.1080/10549811.2020.1841009},
   publisher={Taylor \& Francis}
 }
 
@@ -154,6 +162,7 @@
   number={2},
   pages={e00218--21},
   year={2021},
+  doi={10.1101/2020.10.21.348573},
   publisher={Am Soc Microbiol}
 }
 
@@ -164,6 +173,7 @@
   volume={415},
   pages={107257},
   year={2021},
+  doi={10.1016/j.jvolgeores.2021.107257},
   publisher={Elsevier}
 }
 
@@ -173,6 +183,7 @@
   journal={Physica A: Statistical Mechanics and its Applications},
   pages={126764},
   year={2021},
+  doi={10.1016/j.physa.2021.126764},
   publisher={Elsevier}
 }
 
@@ -180,6 +191,7 @@
   title={Characterizing Stand and Biomass Tables from Diameter Distribution Models: A Case Study for Mixed-Hardwood Forests in Eastern Tennessee, USA},
   author={Yang, Sheng-I and Cao, Quang V and Shoch, David T and Johnson, Trisha},
   journal={Forest Science},
+  doi={10.1093/forsci/fxab051},
   year={2021}
 }
 
@@ -187,5 +199,6 @@
   title={Biological Pattern Discovery with R: Machine Learning Approaches},
   author={Yang, Zheng Rong},
   year={2021},
+  doi={http://dx.doi.org/10.1142/12366},
   publisher={World Scientific}
 }
