+++ paper/paper.bib
@@ -6,17 +6,8 @@
   number={1},
   pages={1--22},
   year={1977},
+  doi={10.1111/j.2517-6161.1977.tb01600.x},
   publisher={Wiley Online Library}
-}
-
-@article{steele2010,
-  title={Performance of Bayesian model selection criteria for Gaussian mixture models},
-  author={Steele, Russell J and Raftery, Adrian E},
-  journal={Frontiers of statistical decision making and bayesian analysis},
-  volume={2},
-  pages={113--130},
-  year={2010},
-  publisher={Citeseer}
 }
 
 @article{feng1996,
@@ -61,6 +52,7 @@
   journal={Journal of statistical software},
   volume={32},
   pages={1--29},
+  doi={10.18637/jss.v032.i06},
   year={2010}
 }
 
@@ -106,6 +98,7 @@
   volume={40},
   number={8},
   pages={1--18},
+  doi={10.18637/jss.v040.i08},
   year={2011}
 }  
 
@@ -117,6 +110,7 @@
   number={2},
   pages={180--185},
   year={2011},
+  doi={10.1002/wics.147},
   publisher={Wiley Online Library}
 }
 
