### 2021-12-29 #4028 @GaryBAYLOR
Thanks! @arfon @fboehm I just added DOIs to all references.
