### 2021-12-31 #4028 @GaryBAYLOR
Hi @soodoku, Thanks for the question. The main features in mixR but not in mixtools package are
(a) `mixR` can fit mixtures models for both raw data and binned data, while `mixtools` is only for raw data (which might not be a very precise summary as `mixtools` provides functions to analyze discrete data such as Poisson data, multinomial data, as well as allowing censored data. But since mixR focuses on mixtures of continuous distributions, e.g., Gaussian, Weibull, Log-normal, Gamma, if we restrict comparison to the mixtures models of continuous distributions then `mixtools` uses raw data and for some mixtures allowing right censored data, as model inputs)

(b) `mixR` uses a single interface function (_mixfit_) for model fitting for both types of data and all distributions, while `mixtools` provides separate functions for modeling different distributions, such as _normalmixEM_, _normalmixEM2comp_, _gammamixEM_, _weibullRMM_SEM_, etc.

(c) Other features such as
      - `mixR` provides mixtures of log-normal
      - `mixR` improves the visualizations using `ggplot2` with flexibility to customize the output to easily produce publication-quality plots
      - `mixR` provides functions to convert parameters between shape-scale format to mean-variance format for Weibull and Gamma.

In summary, (a) is the primary contribution of `mixR` compared to `mixtools`, (b) is the implementation that simplifies how to use the package, (c) are handy features that users often need.

NOTES:
The raw data refers to numeric data x_1, x_2, ..., x_n. The binned data (sometimes called grouped data), is defined as (a_1, b_1, k_1), (a_2, b_2, k_2), ... in which a_i and b_i are cutoff points that form an interval, and k_i is the count of data points that fall in (a_i, b_i). This is the definition in univariate case but the definition can be extended to multivariate data.

Please let me know any other questions for you to better understand and evaluate `mixR`.

Thank you!
Youjiao
