### 2022-01-02 #4031 @whedon
[REVIEW]: mixR: An R package for Finite Mixture Modeling for Both Raw and Binned Data
**Submitting author:** @GaryBAYLOR (<a href="http://orcid.org/0000-0003-0519-9605">Youjiao Yu</a>)
**Repository:** <a href="https://github.com/GaryBAYLOR/mixR.git" target ="_blank">https://github.com/GaryBAYLOR/mixR.git</a>
**Version:** v0.2.1
