### 2022-01-02 GaryBAYLOR/mixR#17 @soodoku
contributor guidelines
JOSS expects community guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines

check this out: https://www.contributor-covenant.org/version/1/0/0/code-of-conduct/

here's how i link to it in my readme: https://github.com/soodoku/tuber#contributor-code-of-conduct
