### 2022-01-02 GaryBAYLOR/mixR#18 @soodoku
statement of need within readme/documentation
I think you have an ok starting point around 'statement of need' in your paper but we JOSS also evaluates if you provide that in your documentation for the software.

I recommend adding a crisp #knowwhy in your readme to orient the users and help guide them to other places if there needs are different.
