### 2022-01-02 8e42d37
Update README.md
Based on two issues raised by soodoku
