### 2022-01-02 GaryBAYLOR/mixR#18 @GaryBAYLOR
Hi @soodoku Thanks for the comment. I have added a _Why mixR ?_ section at the beginning of README.md. Please let me know if it looks good to you or needs further improvement.
