### 2022-01-02 GaryBAYLOR/mixR#17 @GaryBAYLOR
@soodoku Thanks for the information. I have added this section in README.md
