### 2022-01-02 GaryBAYLOR/mixR#20 @soodoku
model object
hey trying some of the code in the readme

```
# Selecting the best g for Normal mixture model
s_normal = select(x2, ncomp = 2:6)
```
It prints out: "The final model: normal mixture (equal variance) with 4 components"

I did
```
str(s_normal)
```
```
List of 5
 $ ncomp    : int [1:10] 2 2 3 3 4 4 5 5 6 6
 $ equal.var: chr [1:10] "Y" "N" "Y" "N" ...
 $ bic      : num [1:10] -387 -387 -403 -443 -445 ...
 $ best     : chr [1:10] " " " " " " " " ...
 $ family   : chr "normal"
 - attr(*, "class")= chr "selectEM"
```

I typed
```s_normal$best
 [1] " " " " " " " " "*" " " " " " " " " " "
```

Is that expected? Why?
