### 2022-01-02 GaryBAYLOR/mixR#21 @soodoku
has bic been consistently implemented?
when i do

```
mod1 = mixfit(x1, ncomp = 2)
```

```
mod1$bic
[1] 4238.221
```

with

```
s_normal = select(x2, ncomp = 2:6)
```

````
s_normal$bic
 [1] -386.7280 -387.1504 -403.0483 -443.2730 -445.2417 -440.1219 -431.4261 -426.6852 -440.6210 -408.1540
```
