### 2022-01-02 GaryBAYLOR/mixR#22 @soodoku
documenting 'ev'
```
?b.test
```

```
a logical value indicating whether the variance of each component should be the same or not (default FALSE). ev is ignored for other family members.
```

the statement `ev is ignored for other family members' is not clear. need to provide default family (normal) somewhere?
