### 2022-01-02 GaryBAYLOR/mixR#23 @soodoku
more helpful error would be useful
```
b1 = bs.test(x1, ncomp = c(2, 3), B = 100, max_iter = 1)
Error in if (any(c(n < 0, pi < 0, sd < 0))) { :
  missing value where TRUE/FALSE needed
```
