### 2022-01-02 GaryBAYLOR/mixR#25 @soodoku
missing error message?
i tried

```
b1 = bs.test(x1, ncomp = c(3, 2), B = 100, max_iter = 2)
```

when documentation =

```
<html>
<body>

a vector of two positive integers specifying the number of components of the mixture model under the null and alternative hypothesis. The first integer should be smaller than the second one. The default value is c(1, 2).
--

</body>
</html>
```

should we throw a warning?
