### 2022-01-02 GaryBAYLOR/mixR#26 @soodoku
test
one kind of test that would be cool = reproduce numbers from another package like mixtools. it may not make it to automated tests but useful to have it somewhere in the docs.
