### 2022-01-02 GaryBAYLOR/mixR#21 @GaryBAYLOR
Hi @soodoku, the BIC is consistently implemented. The reason you see different BIC values is because `mixfit` and `select` returns objects of different classes. `select` is not designed to return the same object as `mixfit` as the former is to compare and visualize candidate models using BIC.
```r
> class(mod1)
[1] "mixfitEM"
> class(s_normal)
[1] "selectEM"
```
`mod1$bic` returns the BIC of the fitted model (scalar), and `s_normal$bic` returns a vector of BIC values for the mixture models fitted and compared with different `ncomp`. In the above example,  when running `select(x2, ncomp = 2:6)`, what happed behind the stage is ten Gaussian mixture models are fitted (ncomp = 2, 3, 4, 5, 6 and for each ncomp, two GMMs are fitted with equal variance and unequal variances).

Actually, if you run `select(x1, ncomp = 2:6)` you will see the second row contains BIC that is the same as `mod1$bic`
```r
> s_normal = select(x1, ncomp = 2:6)
The final model: normal mixture (unequal variances) with 2 components
> s_normal
   ncomp equal.var      bic best
1      2         Y 4309.223
2      2         N 4238.221    *
3      3         Y 4323.038
4      3         N 4256.698
5      4         Y 4272.660
6      4         N 4275.925
7      5         Y 4264.199
8      5         N 4291.169
9      6         Y 4278.018
10     6         N 4311.364
```
Hope this helps!
