### 2022-01-02 GaryBAYLOR/mixR#20 @GaryBAYLOR
Hi @soodoku Thanks for the question. The object of `s_normal` is supposed to be viewed in tabular format shown as below but not extracted independently. I put each column of the table as separate list items so `s_normal$best` is only meaning when the whole `s_normal` is printed. But this is a place I can improve to avoid confusion (to store the table as a whole list item).
```r
> s_normal
   ncomp equal.var      bic best
1      2         Y 4309.223
2      2         N 4238.221    *
3      3         Y 4323.038
4      3         N 4256.698
5      4         Y 4272.660
6      4         N 4275.925
7      5         Y 4264.199
8      5         N 4291.169
9      6         Y 4278.018
10     6         N 4311.364
```
Does this make sense to you? If so I can go ahead and make this change to the function definition of `select` and the corresponding function description in the manual.
