### 2022-01-02 GaryBAYLOR/mixR#21 @soodoku
Agree on all points. I was worried, it is positive in one place and
negative in another. If all good, close the issue.

On Sun, Jan 2, 2022, 2:00 PM Youjiao (Gary) Yu ***@***.***>
wrote:

> Hi @soodoku <https://github.com/soodoku>, BIC could be positive or
> negative depending on the data and model complexity.
>
> I followed the definition of BIC to implement it in the package: BIC = k *
> log(n) - 2 * log(L), in whcih k is the total number of parameters in the
> model, n is the size of data, and log(L) is the log-likelihood (see
> Wikipage <https://en.wikipedia.org/wiki/Bayesian_information_criterion>
> reference).
>
> My implementation is here
> <https://github.com/GaryBAYLOR/mixR/blob/e1a0b8c6dede01b9d8f64b74f8f0bfe84ab6c32f/R/normalEM.R#L42>,
> which is consistent with the definition
> <https://stat.ethz.ch/R-manual/R-devel/library/stats/html/AIC.html> in
> the stats package that comes with base R, in which it describes the
> definition of BIC and says that *"When comparing models fitted by maximum
> likelihood to the same data, the smaller the AIC or BIC, the better the
> fit"*.
>
> Interestingly, some other mixture model packages have different
> implementation for BIC. For example, in mclust the BIC is the BIC defined
> as 2 * log(L) - k * log(n), the opposite of the usual definition. As a
> result, by running the following code we can see very close BIC for the
> best model in mixR and mclust but the values of BIC are in opposite sign.
> For packages like mclust, it would be the larger the AIC or BIC, the
> better the fit.
>
> # library(mixR)
> # library(mclust)
> set.seed(102)
> x1 = rmixnormal(1000, c(0.3, 0.7), c(-2, 3), c(2, 1))
> x2 = rmixweibull(1000, c(0.4, 0.6), c(0.6, 1.3), c(0.1, 0.1))
>
> mod_select_mixR_1 = select(x1, ncomp = 2:6)
> mod_select_mclust_1 = mclustBIC(x1, G = 2:6)
> mod_select_mixR_1
> mod_select_mclust_1
>
> mod_select_mixR_2 = select(x2, ncomp = 2:6)
> mod_select_mclust_2 = mclustBIC(x2, G = 2:6)
> mod_select_mixR_2
> mod_select_mclust_2
>
> Please let me know if I explains clearly and it makes sense to you.
> Thanks!
>
