### 2022-01-02 GaryBAYLOR/mixR#21 @GaryBAYLOR
Thanks @soodoku, I am pretty confident about the implementation of BIC in `mixR` now. An extra point about why BIC is sometimes positive and some other times negative is: for continuous distributions, the likelihood is the sum of log(density) for all data points.  The density could be between 0 and 1 and close to zero (like the density for `x1`), in which case log(density) is negative, and multiple the sum by -2 would lead to a positive BIC. The density could also be well above 1 (for a lot of data points in `x2`), which may produce negative BIC when taking log transform on the density and them multiply by -2.

I will go ahead and close the issue. Thank you for bringing this up and it is an interesting discussion.
