+++ DESCRIPTION
@@ -12,7 +12,7 @@ LazyData: TRUE
 Depends: R(>= 3.5.0)
 Imports: ggplot2 (>= 3.3.3), graphics, Rcpp (>= 1.0.6), stats
 LinkingTo: Rcpp
-RoxygenNote: 7.1.1
+RoxygenNote: 7.1.2
 NeedsCompilation: yes
 Author: Youjiao Yu [aut, cre]
 Repository: CRAN
