### 2022-01-03 GaryBAYLOR/mixR#22 @GaryBAYLOR
Hi @soodoku, this issue is now fixed  in [mixR/R/bs.test.R](https://github.com/GaryBAYLOR/mixR/blob/38b293b34b33ee821d6e85861b2e94a282dbccca/R/bs.test.R#L19) and the corresponding manual file [mixR/man/bs.test.Rd](https://github.com/GaryBAYLOR/mixR/blob/38b293b34b33ee821d6e85861b2e94a282dbccca/man/bs.test.Rd#L33) is also updated accordingly. Thanks!
