+++ R/rmixgamma.R
@@ -22,14 +22,9 @@
 #'
 #' @export
 rmixgamma <- function(n, pi, mu, sd) {
-  if(!(is.numeric(n) || is.numeric(pi) || is.numeric(mu) || is.numeric(sd)))
-  {stop("numeric inputs are required.")}
-  if(length(pi) != length(mu) || (length(pi) != length(sd)))
-  {stop("'pi', 'mu' and 'sd' should have the same length.")}
-  if(any(c(n < 0, pi < 0, sd < 0)))
-  {stop("positive 'n', 'pi' and 'sd' are required.")}
-  if(length(n) > 1) stop("'n' should be a positive integer.")
-
+  check_inputs_for_rmix(n, pi, mu, sd)
+  if(any(mu <= 0)) stop("'mu' should be positive!")
+  
 	n0 <- rmultinom(1, n, pi)[, 1]
 	pars <- to_shape_rate_gamma(mu, sd)
 	alpha <- pars[[1]]
