### 2022-01-03 GaryBAYLOR/mixR#23 @GaryBAYLOR
Hi @soodoku,
The error message you displayed is caused by setting `max_iter` too small, and backtracking revealed that a the function `rmixnormal` got an invalid input
```r
rmixnormal(1000, c(NaN, NaN), c(0, 0), c(0, 0))
```
To fix the issue by showing more informative error message I updated two areas:
(1) restrict `max_iter` to be at least 200 as if this number is too low, EM algorithm has a high risk of not converging.
After this fix, we can get the following error message
```r
b1 = bs.test(x1, ncomp = c(2, 3), B = 100, max_iter = 1)
 Error in normalEM(x = x1, ncomp = 2, max_iter = 1, ev = FALSE, mstep.method = "bisection",  :
  'max_iter' is too small! Recommend to leave it as default or set it at least 200 to reduce the risk of EM algorithm not converging.
```

(2) adding more complete error checking for random number generation functions, so when we enter an invalid input we see the following error message
```r
rmixnormal(1000, c(NaN, NaN), c(0, 0), c(0, 0))
 Error in check_inputs_for_rmix(n, pi, mu, sd) :
  'n', 'pi', 'mu' and 'sd' can not contain missing values!
```

Here is the [pull request](https://github.com/GaryBAYLOR/mixR/pull/29). Please let me know if this looks good so I can merge it to main branch.

Thanks a lot!
