### 2022-01-03 GaryBAYLOR/mixR#25 @GaryBAYLOR
Hi @soodoku, I just fixed this by throwing an error message when the first number in `ncomp` is greater than the second one (instead of a warning for the purpose of forcing users to use correct input). I added a few more error checking for `ncomp`. Now it looks like
```r
> b1 = bs.test(x1, ncomp = c(3, 2), B = 100, max_iter = 2)
Error in bs.test(x1, ncomp = c(3, 2), B = 100, max_iter = 2) :
  ncomp[1] must be less than ncomp[2] after coercion to integers!

> b1 = bs.test(x1, ncomp = c(0, 2), B = 100, max_iter = 2)
Error in bs.test(x1, ncomp = c(0, 2), B = 100, max_iter = 2) :
  'ncomp' must be positive!

> b1 = bs.test(x1, ncomp = c(NA, 2), B = 100, max_iter = 2)
Error in bs.test(x1, ncomp = c(NA, 2), B = 100, max_iter = 2) :
  'ncomp' must not contain missing value!

> b1 = bs.test(x1, ncomp = c('a', 2), B = 100, max_iter = 2)
Error in bs.test(x1, ncomp = c("a", 2), B = 100, max_iter = 2) :
  'ncomp' must be a numeric vector!

> b1 = bs.test(x1, ncomp = c(FALSE, 2), B = 100, max_iter = 2)
Error in bs.test(x1, ncomp = c(FALSE, 2), B = 100, max_iter = 2) :
  'ncomp' must be positive!

> b1 = bs.test(x1, ncomp = c(1, 100), B = 100, max_iter = 2)
Error in bs.test(x1, ncomp = c(1, 100), B = 100, max_iter = 2) :
  'ncomp' larger than 20 is currently not supported!

```
