### 2022-01-03 GaryBAYLOR/mixR#26 @GaryBAYLOR
Hi @soodoku,
I can create an R vignette for comparing a few similar packages including `mixR`, `mixtools`, `mixdist`, and `mclust`, etc. Since each of them focuses on different aspects of mixture model problems, the comparison will be for things that both `mixR` and X can do when we compare `mixR` with X.

In this way we are not adding comparison information into the current tests (the unit tests fulfilled by `testthat` package) but instead into a vignette. Users can also get access to it quickly from CRAN.

Please let me if this sounds like a good idea to you. If so I can start working on it (may take a few days).

Thanks!
