### 2022-01-03 GaryBAYLOR/mixR#26 @soodoku
the thing that i am not checking but worried about = correctness.

one way to get at that = make sure results agree from diff. software as there are already implementations.

i don't think a matrix of comparison is useful otherwise except in a way where you do product comparison via a table.
