### 2022-01-04 64fd9d9
Adding tests for rmix functions
