### 2022-01-04 #4031 @welch16
Hey @fboehm, this is my first time reviewing for JOSS, so I was wondering if I can comment on the package vignette and such?
