### 2022-01-04 4463a22
Create test_compare_with_mixtools.R
