### 2022-01-04 GaryBAYLOR/mixR#26 @GaryBAYLOR
Hi @soodoku,

I added some unit tests for testing the correctness of `mixR` by comparing the results from `mixR` and `mixtools` (here is the [pull request](https://github.com/GaryBAYLOR/mixR/pull/31/files)). The tests all passed but only cover Normal and Gamma as I found the Weibull mixture is not that reliable in mixtools when not providing initial values. Log-normal is not supported by `mixtools`.

Another way to test the correctness of mixture model fitting is by simulation.
* Generating some random data from a mixture distribution with known parameters
* Fitting a mixture model to the data with the same number of mixture components with true data
* Checking if the fitted parameters are close to the true parameters

Here is an example,
```r
n = 5000 # sample size
lambda=c(0.7, 0.3)
shape <- c(1, 5); scale <- c(2,20) # model parameters
set.seed(321)
x <- rweibullmix(n, lambda, shape, scale) # iid ~ weibull mixture

t1 <- Sys.time()
mod_mixtools <- mixtools::weibullRMM_SEM(x, d = rep(1, length(x)), k = 2, maxit = 500)
t2 <- Sys.time()
mod_mixR <- mixfit(x, ncomp = 2, family = 'weibull')
t3 <- Sys.time()
```
The comparison of fitted results is as follows. You can see that the fitted parameters using `mixtools` is way off.
```r
> # time used by mixtools
> t2 - t1
Time difference of 20.7195 secs
> # time used by mixR
> t3 - t2
Time difference of 0.6110721 secs
>
> # fitted parameter: proportion
> mod_mixtools$lambda
[1] 0.4794032 0.5205968
> mod_mixR$pi
[1] 0.6933603 0.3066397
>
> # fitted parameter: shape
> mod_mixtools$shape
[1] 0.7142412 0.7525246
> mod_mixR$k
[1] 1.012320 4.976074
>
> # fitted parameter: scale
> mod_mixtools$scale
[1] 5.217042 6.340802
> mod_mixR$lambda
[1]  2.008306 20.066505
>
> # log-likelihood
> mod_mixtools$loglik
[1] -14306.31
> mod_mixR$loglik
[1] -13149.67
```

If we provide good initial values, `mixtools` gives similar results as `mixR`, both of which are close to true parameters.
```r
t1 <- Sys.time()
mod_mixtools <- weibullRMM_SEM(x, d = rep(1, length(x)), k = 2,
                               lambda = c(0.8, 0.2), shape = c(1.5, 6), scale = c(3, 15),
                               maxit = 500)
t2 <- Sys.time()
mod_mixR <- mixfit(x, ncomp = 2, family = 'weibull')
t3 <- Sys.time()

> # time used by mixtools
> t2 - t1
Time difference of 20.45888 secs
> # time used by mixR
> t3 - t2
Time difference of 0.6568961 secs
>
> # fitted parameter: proportion
> mod_mixtools$lambda
[1] 0.6937048 0.3062952
> mod_mixR$pi
[1] 0.6933603 0.3066397
>
> # fitted parameter: shape
> mod_mixtools$shape
[1] 1.01434 4.98554
> mod_mixR$k
[1] 1.012320 4.976074
>
> # fitted parameter: scale
> mod_mixtools$scale
[1]  2.007941 20.058944
> mod_mixR$lambda
[1]  2.008306 20.066505
>
> # log-likelihood
> mod_mixtools$loglik
[1] -13149.69
> mod_mixR$loglik
[1] -13149.67
```

A third way to validate the correctness is by visualizing the fitted results and overlay them over the histogram of the data. If the model fitting is good then the the fitted density curve should match with the shape of the histogram.

Hope this makes sense to you!

Thank you
