### 2022-01-04 GaryBAYLOR/mixR#26 @soodoku
thanks, @GaryBAYLOR! this provides you another little point for mixR.

you are gtg from my end now. one down re. JOSS reviews.
