### 2022-01-04 GaryBAYLOR/mixR#26 @GaryBAYLOR
Thanks @soodoku. I really appreciate all your comments and discussions from which I also got a deeper understanding of mixture models.
