### 2022-01-04 #4031 @fboehm
@soodoku and @welch16 - Thanks so much for your work on the reviews! Yes, @welch16 - please do comment on the vignettes and other features of the package as needed. Please feel free to post comments in this thread and to open issues in the package repository as you work through the package. Please don't hesitate to ask me any questions that arise. Thanks again!
