### 2022-01-05 GaryBAYLOR/mixR#32 @welch16
Binned data format
I think that the one of the main contributions of this package is that allows the user to work with binned data. However, the examples in the vignette that utilize binned data are from simulated data that was later binned using the `mixR::bin` function which may confuse an inexperienced user.

I consider that the vignette and the README file could be improved with an explicit example of the binned data matrix expected by the `mixfit` function that wasn't generated with `mixR::bin`
