### 2022-01-05 #4031 @welch16
I create an issue here: https://github.com/GaryBAYLOR/mixR/issues/32. But except that, I think everything looks good
