### 2022-01-05 GaryBAYLOR/mixR#32 @GaryBAYLOR
Hi @welch16,
I have added two examples in the vignette for fitting mixtures models with binned data which is not binned using `mixR::bin`. One is `Stamp2` data in mixR pacakge. The other one is `cassie` data from `mixdist` package. Please refer to Section 2.4 in the [vignette](https://github.com/GaryBAYLOR/mixR/blob/issue_32/vignettes/intro-to-mixr.pdf) for details.  I also updated the README.md file.

The vignette and other changes are currently in Github but will be updated to CRAN in the next release.

Please let me know it looks good to you now.

Thanks!
