### 2022-01-05 #4031 @GaryBAYLOR
Hi @fboehm,
Here are links to all issues raised by @soodoku.
Contributor Guidelines: https://github.com/GaryBAYLOR/mixR/issues/17
statement of need within readme: https://github.com/GaryBAYLOR/mixR/issues/18
model object: https://github.com/GaryBAYLOR/mixR/issues/20
has bic been consistently implemented? https://github.com/GaryBAYLOR/mixR/issues/21
documenting 'ev': https://github.com/GaryBAYLOR/mixR/issues/22
more helpful error would be useful: https://github.com/GaryBAYLOR/mixR/issues/23
missing error message? https://github.com/GaryBAYLOR/mixR/issues/25
more unit tests: https://github.com/GaryBAYLOR/mixR/issues/26

Thanks!
