### 2022-01-11 #4031 @fboehm
Thanks, @welch16 and @soodoku for your constructive reviews. Thanks, @GaryBAYLOR - there are a few remaining steps before publication. Let me proofread the manuscript in the next two days. I'll then provide instructions for the remaining steps - namely, making an archive and a new release... but don't do those steps until I get back to you after the proofreading. Thanks again!
