### 2022-01-12 #4031 @fboehm
@GaryBAYLOR - here are a few minor edits needed for paper.md:

1. Line 29 (of the pdf): ggplot should be ggplot2. Also, be sure to typeset ggplot2 as you did the other package names.
2. Line 48 (of the pdf): "binning is it" should be "binning is that it"
3. line 94 (of the pdf): "better than it with" should be "better than that with"
4. Figure 2 (top left and top right) : Do you want to change the axis tick labels to be only on the integers? Since you're considering only integer values? Right now, they have 0.5 increments.
