### 2022-01-12 #4031 @fboehm
@GaryBAYLOR - The references section of the pdf needs a few edits:

1. In Eddelbuettel, et al., capitalize R and C (in C++).
2. in Benaglia, et al., capitalize R.
3. Korne, et al. has the wrong doi: use https://dx.doi.org/10.1128/mSphere.00218-21. The currently listed doi directs to the preprint.
4. Leisch (2004) needs a capital R. Also, the "m" in FlexMix needs to be capitalized.
5. Scrucca, et al. needs a lower case "m" in "mclust", even though it is the first letter of the title.
6. Sylvestre, et al. needs capital "E." and lower-case "coli".
7. Yang, et al. needs a capital "T" in "Tennessee".
8. Yang (2021) needs a capital "R".
