### 2022-01-12 7d70686
Updating paper.md, paper.bib and plot.selectEM.R
