+++ paper/paper.bib
@@ -23,7 +23,7 @@
 }
 
 @article{mclachlan1987,
-  title={On bootstrapping the likelihood ratio test statistic for the number of components in a normal mixture},
+  title={On bootstrapping the likelihood ratio test statistic for the number of components in a {N}ormal mixture},
   author={McLachlan, Geoffrey J},
   journal={Journal of the Royal Statistical Society: Series C (Applied Statistics)},
   volume={36},
@@ -35,7 +35,7 @@
 }
 
 @article{yu2019,
-  title={Bootstrap likelihood ratio test for Weibull mixture models fitted to grouped data},
+  title={Bootstrap likelihood ratio test for {W}eibull mixture models fitted to grouped data},
   author={Yu, Youjiao and Harvill, Jane L},
   journal={Communications in Statistics-Theory and Methods},
   volume={48},
@@ -47,7 +47,7 @@
 }
 
 @article{mixtools,
-  title={mixtools: an R package for analyzing mixture models},
+  title={{m}ixtools: an {R} package for analyzing mixture models},
   author={Benaglia, Tatiana and Chauveau, Didier and Hunter, David R and Young, Derek S},
   journal={Journal of statistical software},
   volume={32},
@@ -57,7 +57,7 @@
 }
 
 @article{mclust,
-  title={mclust 5: clustering, classification and density estimation using Gaussian finite mixture models},
+  title={{m}clust 5: clustering, classification and density estimation using {G}aussian finite mixture models},
   author={Scrucca, Luca and Fop, Michael and Murphy, T Brendan and Raftery, Adrian E},
   journal={The R journal},
   volume={8},
@@ -69,7 +69,7 @@
 }
 
 @article{flexmix,
-  title={Flexmix: A general framework for finite mixture models and latent glass regression in R},
+  title={Flex{M}ix: A general framework for finite mixture models and latent glass regression in R},
   author={Leisch, Friedrich},
   doi={10.18637/jss.v011.i08},
   year={2004}
@@ -92,7 +92,7 @@
 }
   
 @article{rcpp,
-  title={Rcpp: Seamless R and C++ integration},
+  title={Rcpp: Seamless {R} and {C}++ integration},
   author={Eddelbuettel, Dirk and Fran{\c{c}}ois, Romain and Allaire, J and Ushey, Kevin and Kou, Qiang and Russel, N and Chambers, John and Bates, D},
   journal={Journal of statistical software},
   volume={40},
@@ -128,7 +128,7 @@
 }
 
 @article{sylvestre2020,
-  title={Can routine monitoring of E. coli fully account for peak event concentrations at drinking water intakes in agricultural and urban rivers?},
+  title={Can routine monitoring of {E}. {c}oli fully account for peak event concentrations at drinking water intakes in agricultural and urban rivers?},
   author={Sylvestre, {\'E}mile and Burnet, Jean-Baptiste and Smeets, Patrick and Medema, Gertjan and Pr{\'e}vost, Mich{\`e}le and Dorner, Sarah},
   journal={Water research},
   volume={170},
@@ -156,7 +156,7 @@
   number={2},
   pages={e00218--21},
   year={2021},
-  doi={10.1101/2020.10.21.348573},
+  doi={10.1128/mSphere.00218-21},
   publisher={Am Soc Microbiol}
 }
 
@@ -190,7 +190,7 @@
 }
 
 @misc{yang2021bio,
-  title={Biological Pattern Discovery with R: Machine Learning Approaches},
+  title={Biological Pattern Discovery with {R}: Machine Learning Approaches},
   author={Yang, Zheng Rong},
   year={2021},
   doi={10.1142/12366},
