+++ paper/paper.bib
@@ -69,7 +69,7 @@
 }
 
 @article{flexmix,
-  title={Flex{M}ix: A general framework for finite mixture models and latent glass regression in R},
+  title={Flex{M}ix: A general framework for finite mixture models and latent glass regression in {R}},
   author={Leisch, Friedrich},
   doi={10.18637/jss.v011.i08},
   year={2004}
@@ -182,7 +182,7 @@
 }
 
 @article{yang2021,
-  title={Characterizing Stand and Biomass Tables from Diameter Distribution Models: A Case Study for Mixed-Hardwood Forests in Eastern Tennessee, USA},
+  title={Characterizing Stand and Biomass Tables from Diameter Distribution Models: A Case Study for Mixed-Hardwood Forests in Eastern {T}ennessee, USA},
   author={Yang, Sheng-I and Cao, Quang V and Shoch, David T and Johnson, Trisha},
   journal={Forest Science},
   doi={10.1093/forsci/fxab051},
