### 2022-01-12 #4031 @GaryBAYLOR
Hi @fboehm, thanks for your proofreading and detailed comments. I fixed all issues about the text, plots and the references.
* For the plots (the first two in Figure 2), I updated the paper as well as the R function plot.selectEM() in the package to only plot integer values on x axis.
* For issues in the references, I just realize that JOSS compiler automatically changes the first letter of the first word in the paper title to uppercase, and all other letters to lowercase (if not followed by .). So I manually forced letters to be uppercase or lowercase by using {A} or {a}.

Please let me know if you still have other comments. And again, I appreciate your help in improving the quality of the paper.
