### 2022-01-12 #4031 @fboehm
Thanks so much, @GaryBAYLOR !

Please create an archive, with a service like zenodo, and a new git release of your package. Please be sure that the title and authors for the archive match exactly those of the paper.md. Once you've done those two tasks, please report here the version number for the release and the doi for the archive. Let me know if you have any questions about this. Thanks again!
