### 2022-01-12 #4031 @GaryBAYLOR
Hi @fboehm,
I have created a new release of mixR on Github with version number v0.2.1 (see [here](https://github.com/GaryBAYLOR/mixR/tree/v0.2.1))
I have created an archive on Zenodo with DOI: 10.5281/zenodo.5842096 (see [here](https://zenodo.org/record/5842096#.Yd8x3xPMKFi)).

Can you help check if both the git release and zenodo archive are in correct format required by JOSS? Thanks!
Youjiao
